requirements.txt
streamlit>=1.20.0
langchain>=0.0.278
openai>=0.27.0
qdrant-client>=1.0.0
langchain-qdrant>=0.0.5
python-dotenv>=0.21.0
beautifulsoup4>=4.12.0
requests>=2.28.0
tiktoken>=0.4.0